backend/requirements.txt
flask
flask-cors
flask-bcrypt
flask-jwt-extended
numpy
opencv-python-headless
mediapipe
pandas
scipy
gunicorn
pymongo
python-dotenv
mangum
